+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 dev
 ---
 
+* `Fossil.Mixins.Layoutable`: Does not execute template function anymore, assume
+  it is a compiled template. Method is no more evaluated in Layoutable context.
+* `Fossil.Mixins.Layoutable`: Backbone.View used as template behave like native
+  views and are append using `setElement`.
 * No more needs for application parameter on Modules.
   It is set on application connection. Introduces the `connect` event for
   modules.